yfinance-mcp-server/pyproject.toml
[project]
name = "yfinance-mcp-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "yfinance (>=0.2.55,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mcp (>=1.6.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
